+++ spec-main/api-browser-window-spec.ts
@@ -2,10 +2,11 @@ import * as chai from 'chai'
 import * as chaiAsPromised from 'chai-as-promised'
 import * as path from 'path'
 import * as fs from 'fs'
+import * as os from 'os'
 import * as qs from 'querystring'
 import * as http from 'http'
 import { AddressInfo } from 'net'
-import { app, BrowserWindow, ipcMain, OnBeforeSendHeadersListenerDetails, screen, webContents } from 'electron'
+import { app, BrowserWindow, BrowserView, ipcMain, OnBeforeSendHeadersListenerDetails, screen, webContents } from 'electron'
 import { emittedOnce } from './events-helpers';
 import { closeWindow } from './window-helpers';
 
@@ -1195,4 +1196,174 @@ describe('BrowserWindow module', () => {
       w.webContents.openDevTools()
     })
   })
+
+  describe('BrowserWindow.fromBrowserView(browserView)', () => {
+    afterEach(closeAllWindows)
+
+    it('returns the window with the browserView', () => {
+      const w = new BrowserWindow({ show: false })
+      const bv = new BrowserView
+      w.setBrowserView(bv)
+      expect(BrowserWindow.fromBrowserView(bv)!.id).to.equal(w.id)
+    })
+
+    it('returns undefined if not attached', () => {
+      const bv = new BrowserView
+      expect(BrowserWindow.fromBrowserView(bv)).to.be.null('BrowserWindow associated with bv')
+    })
+  })
+
+  describe('BrowserWindow.setOpacity(opacity)', () => {
+    afterEach(closeAllWindows)
+    it('make window with initial opacity', () => {
+      const w = new BrowserWindow({ show: false, opacity: 0.5 })
+      expect(w.getOpacity()).to.equal(0.5)
+    })
+    it('allows setting the opacity', () => {
+      const w = new BrowserWindow({ show: false })
+      expect(() => {
+        w.setOpacity(0.0)
+        expect(w.getOpacity()).to.equal(0.0)
+        w.setOpacity(0.5)
+        expect(w.getOpacity()).to.equal(0.5)
+        w.setOpacity(1.0)
+        expect(w.getOpacity()).to.equal(1.0)
+      }).to.not.throw()
+    })
+  })
+
+  describe('BrowserWindow.setShape(rects)', () => {
+    afterEach(closeAllWindows)
+    it('allows setting shape', () => {
+      const w = new BrowserWindow({ show: false })
+      expect(() => {
+        w.setShape([])
+        w.setShape([{ x: 0, y: 0, width: 100, height: 100 }])
+        w.setShape([{ x: 0, y: 0, width: 100, height: 100 }, { x: 0, y: 200, width: 1000, height: 100 }])
+        w.setShape([])
+      }).to.not.throw()
+    })
+  })
+
+  describe('"useContentSize" option', () => {
+    afterEach(closeAllWindows)
+    it('make window created with content size when used', () => {
+      const w = new BrowserWindow({
+        show: false,
+        width: 400,
+        height: 400,
+        useContentSize: true
+      })
+      const contentSize = w.getContentSize()
+      expect(contentSize).to.deep.equal([400, 400])
+    })
+    it('make window created with window size when not used', () => {
+      const w = new BrowserWindow({
+        show: false,
+        width: 400,
+        height: 400,
+      })
+      const size = w.getSize()
+      expect(size).to.deep.equal([400, 400])
+    })
+    it('works for a frameless window', () => {
+      const w = new BrowserWindow({
+        show: false,
+        frame: false,
+        width: 400,
+        height: 400,
+        useContentSize: true
+      })
+      const contentSize = w.getContentSize()
+      expect(contentSize).to.deep.equal([400, 400])
+      const size = w.getSize()
+      expect(size).to.deep.equal([400, 400])
+    })
+  })
+
+  ifdescribe(process.platform === 'darwin' && parseInt(os.release().split('.')[0]) >= 14)('"titleBarStyle" option', () => {
+    afterEach(closeAllWindows)
+    it('creates browser window with hidden title bar', () => {
+      const w = new BrowserWindow({
+        show: false,
+        width: 400,
+        height: 400,
+        titleBarStyle: 'hidden'
+      })
+      const contentSize = w.getContentSize()
+      expect(contentSize).to.deep.equal([400, 400])
+    })
+    it('creates browser window with hidden inset title bar', () => {
+      const w = new BrowserWindow({
+        show: false,
+        width: 400,
+        height: 400,
+        titleBarStyle: 'hiddenInset'
+      })
+      const contentSize = w.getContentSize()
+      expect(contentSize).to.deep.equal([400, 400])
+    })
+  })
+
+  ifdescribe(process.platform !== 'linux')('"enableLargerThanScreen" option', () => {
+    afterEach(closeAllWindows)
+    it('can move the window out of screen', () => {
+      const w = new BrowserWindow({ show: true, enableLargerThanScreen: true })
+      w.setPosition(-10, -10)
+      const after = w.getPosition()
+      expect(after).to.deep.equal([-10, -10])
+    })
+    it('without it, cannot move the window out of screen', () => {
+      const w = new BrowserWindow({ show: true, enableLargerThanScreen: false })
+      w.setPosition(-10, -10)
+      const after = w.getPosition()
+      expect(after[1]).to.be.at.least(0)
+    })
+    it('can set the window larger than screen', () => {
+      const w = new BrowserWindow({ show: true, enableLargerThanScreen: true })
+      const size = screen.getPrimaryDisplay().size
+      size.width += 100
+      size.height += 100
+      w.setSize(size.width, size.height)
+      expectBoundsEqual(w.getSize(), [size.width, size.height])
+    })
+    it('without it, cannot set the window larger than screen', () => {
+      const w = new BrowserWindow({ show: true, enableLargerThanScreen: false })
+      const size = screen.getPrimaryDisplay().size
+      size.width += 100
+      size.height += 100
+      w.setSize(size.width, size.height)
+      expect(w.getSize()[1]).to.at.most(screen.getPrimaryDisplay().size.height)
+    })
+  })
+
+  ifdescribe(process.platform === 'darwin')('"zoomToPageWidth" option', () => {
+    afterEach(closeAllWindows)
+    it('sets the window width to the page width when used', () => {
+      const w = new BrowserWindow({
+        show: false,
+        width: 500,
+        height: 400,
+        zoomToPageWidth: true
+      })
+      w.maximize()
+      expect(w.getSize()[0]).to.equal(500)
+    })
+  })
+
+  describe('"tabbingIdentifier" option', () => {
+    afterEach(closeAllWindows)
+    it('can be set on a window', () => {
+      expect(() => {
+        new BrowserWindow({
+          tabbingIdentifier: 'group1'
+        })
+        new BrowserWindow({
+          tabbingIdentifier: 'group2',
+          frame: false
+        })
+      }).not.to.throw()
+    })
+  })
+
 })
